+++ CranberryAPI/Models/RegisterDto.cs
@@ -14,7 +14,7 @@ public class RegisterDto
 
   public string QuitDate { get; set; }
   public int AvgSmokedDaily { get; set; }
-  public int PricePerPack { get; set; }
+  public float PricePerPack { get; set; }
   public int CigsPerPack { get; set; }
 
 }
